Adding redirect for directories without a / to add the / - the reason is that it simplifies relative URLs.
Consider requesting path/to/file

Where "file" is a directory, old method would load path/to/file/index (if it exists), which might have an
a href='relativefile' expecting that to produce path/to/file/relativefile - but it wouldn't!

So instead we now redirect to path/to/file/

Now that "relativefile" link would work as expected.